+++ masuite-common/src/main/java/dev/masa/masuite/common/models/home/Home.java
@@ -1,8 +1,8 @@
-package dev.masa.masuite.common.models;
+package dev.masa.masuite.common.models.home;
 
 import com.j256.ormlite.field.DataType;
 import com.j256.ormlite.field.DatabaseField;
-import dev.masa.masuite.api.models.warp.IWarp;
+import dev.masa.masuite.api.models.home.IHome;
 import dev.masa.masuite.common.objects.Location;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -16,19 +16,16 @@ import java.util.UUID;
 @NoArgsConstructor()
 @Data()
 @Entity()
-@Table(name = "masuite_warps")
-public class Warp implements IWarp {
+@Table(name = "masuite_homes")
+public class Home implements IHome {
 
     @DatabaseField(dataType = DataType.UUID, generatedId = true, columnName = "uuid")
     private UUID uniqueId;
     @DatabaseField
     private String name;
 
-    @DatabaseField(columnName = "public")
-    private boolean isPublic;
-
-    @DatabaseField
-    private boolean global;
+    @DatabaseField(dataType = DataType.UUID)
+    private UUID owner;
 
     @DatabaseField
     private Double x;
@@ -51,21 +48,10 @@ public class Warp implements IWarp {
     @DatabaseField
     private String world;
 
-    public Warp(String name, Location location, boolean isPublic, boolean isGlobal) {
+    public Home(String name, UUID owner, Location location) {
         this.name = name;
-        this.isPublic = isPublic;
-        this.global = isGlobal;
+        this.owner = owner;
         this.location(location);
-    }
-
-    @Override
-    public boolean isPublic() {
-        return this.isPublic;
-    }
-
-    @Override
-    public boolean isGlobal() {
-        return this.global;
     }
 
     public Location location() {
