+++ masuite-common/src/main/java/dev/masa/masuite/common/models/user/User.java
@@ -1,4 +1,4 @@
-package dev.masa.masuite.common.models;
+package dev.masa.masuite.common.models.user;
 
 import com.j256.ormlite.field.DataType;
 import com.j256.ormlite.field.DatabaseField;
