+++ masuite-velocity/src/main/java/dev/masa/masuite/velocity/listeners/home/DeleteHomeMessageListener.java
@@ -4,8 +4,8 @@ import com.velocitypowered.api.event.Subscribe;
 import com.velocitypowered.api.event.connection.PluginMessageEvent;
 import com.velocitypowered.api.proxy.Player;
 import dev.masa.masuite.api.proxy.listeners.home.IDeleteHomeMessageListener;
-import dev.masa.masuite.common.models.Home;
-import dev.masa.masuite.common.models.User;
+import dev.masa.masuite.common.models.home.Home;
+import dev.masa.masuite.common.models.user.User;
 import dev.masa.masuite.common.objects.MaSuiteMessage;
 import dev.masa.masuite.common.services.MessageService;
 import dev.masa.masuite.velocity.MaSuiteVelocity;
